Saturday morning fun: add whoami section and make portone a slug
