--- src/components/portOne/portOne.scss
@@ -1,100 +1,0 @@
-@use './portOne-variables';
-@use '../../styles/variables';
-
-.portone {
-    width: 100vw;
-    height: 100vh;
-    background-color: portOne-variables.$top-hero-gradient;
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: #fff;
-
-    background: linear-gradient(
-        0deg,
-        portOne-variables.$bottom-hero-gradient 0%,
-        portOne-variables.$top-hero-gradient 30%
-    );
-
-    .heading {
-        text-align: center;
-
-        &__main {
-            line-height: 148px;
-            letter-spacing: -11px;
-            font-size: 184px;
-            font-family: Helvetica, sans-serif;
-            display: flex;
-            flex-flow: column;
-
-            span.italic {
-                font-style: italic;
-            }
-        }
-
-        &__sub {
-            margin: 15px auto;
-            max-width: 700px;
-            font-family: Helvetica, sans-serif;
-            text-transform: uppercase;
-            font-size: 16px;
-            line-height: 18px;
-        }
-    }
-
-    .handlebar {
-        position: fixed;
-        left: 50%;
-        transform: translateX(-50%);
-        bottom: 50px;
-        font-size: 20px;
-        padding: 25px 30px;
-        backdrop-filter: blur(20px);
-        border-radius: 15px;
-        background-color: rgba(0, 0, 0, 0.3);
-        font-family: Helvetica, sans-serif;
-
-        &__inner {
-            position: relative;
-        }
-
-        &__decoration {
-            position: fixed;
-        }
-
-        .decoration {
-            width: 48px;
-            height: 2px;
-            border-radius: 2px;
-            background-color: #fff;
-            transition: width 0.2s, left 0.4s;
-            transition-timing-function: ease-in-out;
-            will-change: left;
-        }
-
-        ul {
-            position: relative;
-            display: flex;
-
-            li {
-                list-style: none;
-                user-select: none;
-
-                button {
-                    background-color: transparent;
-                    color: #fff;
-                    border: none;
-                    padding: 0;
-                    font-size: 18px;
-                    font-weight: bold;
-                    cursor: pointer;
-                }
-
-                &:not(:last-child) {
-                    margin-right: 50px;
-                }
-            }
-        }
-    }
-}
