+++ src/components/portOne/whoami/whoami.scss
@@ -1,0 +1,81 @@
+@use '../portOne-variables';
+@use '../../../styles/variables';
+
+.whoami {
+    background-color: #f5f0f0;
+    min-height: 100vh;
+    color: #000;
+    font-family: Helvetica, sans-serif;
+
+    &__inner {
+        width: 100%;
+        max-width: 60%;
+        margin: 0 auto 0 10%;
+        position: relative;
+        display: flex;
+        flex-flow: column;
+        align-items: flex-end;
+        padding-top: 10vw;
+    }
+
+    h1,
+    h2,
+    p {
+        font-size: 4.7vw;
+        line-height: 5vw;
+    }
+
+    h2 {
+        white-space: nowrap;
+        margin: 2vw 0;
+    }
+
+    &__floater {
+        // position: absolute;
+        margin-top: 20px;
+        font-size: 18px !important;
+        line-height: 22px !important;
+        max-width: 500px;
+        top: 25%;
+        right: -75%;
+        text-align: right;
+    }
+
+    &__weirdo {
+        width: 100%;
+        height: 1070px;
+        background-image: url(../../../images/unsplash_4-Lo8Pkxs_s.png);
+        margin: 20px 0;
+        border-radius: 30px;
+        overflow: hidden;
+    }
+
+    &__links {
+        width: 100%;
+        margin-bottom: 400px;
+
+        ul {
+            display: flex;
+            justify-content: space-between;
+
+            li {
+                font-size: 2vw;
+
+                a {
+                    text-decoration: underline;
+                    cursor: pointer;
+                }
+
+                a,
+                a:visited {
+                    color: #000;
+                }
+
+                span {
+                    font-size: 1vw;
+                    margin-left: -5px;
+                }
+            }
+        }
+    }
+}
