--- src/pages/port-one.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { PortOne } from '../components/portOne/PortOne';
-
-const PersonalSpace = () => {
-    return <PortOne />;
-};
-
-export default PersonalSpace;
