+++ src/pages/port-one/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { PortOne } from '../../components/portOne/index/PortOne';
+
+const PersonalSpace = () => {
+    return <PortOne />;
+};
+
+export default PersonalSpace;
