+++ src/pages/port-one/whoami.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Whoami } from '../../components/portOne/whoami/Whoami';
+
+const PersonalSpace = () => {
+    return <Whoami />;
+};
+
+export default PersonalSpace;
